backend/requirements.txt
flask
mysql-connector-python
python-dotenv
flask-cors  # used for CORS
PyJWT
APScheduler
bcrypt
